+++ R/GetShinyUserRole.R
@@ -7,12 +7,6 @@
 #' @seealso \code{\link{getShinyUserName}}, \code{\link{getShinyUserGroups}},
 #' \code{\link{getShinyUserReshId}}
 #'
-#' @examples
-#' \donttest{
-#' # Requires a valid shiny session object
-#' getShinyUserRole(shinySessionObject)
-#' }
-#'
 #' @export