+++ R/GetUserPhone.R
@@ -15,8 +15,8 @@
 #' @examples
 #' \donttest{
 #' # Requires a valid shiny session object
-#' getUserPhone()
-#' getUserPhone(shinySessionObject)
+#' try(getUserPhone())
+#' try(getUserPhone(shinySessionObject))
 #' }
 #'
 #' @export