+++ R/stats.R
@@ -10,6 +10,13 @@
 #'
 #' @param id Character string shiny module id
 #' @param registryName Character string registry name key
+#' @param type Character string defining data level. One of
+#' \code{c("app", "report")}.
+#' @param log Data frame containing log data (in Rapporteket format)
+#' @param startDate Date object defining start of interval (character
+#' representation "YYYY-MM-DD")
+#' @param endDate Date object defining end of interval (character representation
+#' "YYYY-MM-DD")
 #'
 #' @return Shiny objects, mostly. Helper functions may return other stuff too.
 #' @name stats
@@ -164,14 +171,14 @@ logFormat <- function(log) {
 
   log <- log %>%
     dplyr::mutate(
-      datetime = as.POSIXct(time),
-      date = as.Date(datetime),
-      time = format(datetime, "%H:%M:%S"),
-      year = lubridate::year(datetime),
-      month = lubridate::month(datetime),
-      day = lubridate::day(datetime),
+      datetime = as.POSIXct(.data$time),
+      date = as.Date(.data$datetime),
+      time = format(.data$datetime, "%H:%M:%S"),
+      year = lubridate::year(.data$datetime),
+      month = lubridate::month(.data$datetime),
+      day = lubridate::day(.data$datetime),
       weekday = lubridate::wday(
-        datetime,
+        .data$datetime,
         week_start = 1,
         abbr = FALSE)
     )