+++ NEWS.md
@@ -1,3 +1,7 @@
+# rapbase 1.16.0
+
+* New functions and shiny modules providing automated reports (subscriptions, dispatchments and bulletins) for registries ([#78](https://github.com/Rapporteket/rapbase/pull/78))
+
 # rapbase 1.15.1
 
 * Allow GitHub API calls by proxy ([#76](https://github.com/Rapporteket/rapbase/pull/76))